refactor: extract components from App.jsx and use Zustand for state management;
